my cyber crime project/ai_backend/requirements.txt
flask
flask-cors
PyMuPDF
ctransformers[cuda]
faiss-cpu
sentence-transformers